Set the json model to null in the OneOf and AnyOf interfaces
The OneOf and AnyOf interfaces added as part of PR #367 broke the html
doc generation.
The interfaces are added as models in the javascript generation with:
```
defs.{{name}} = {{{modelJson}}};
```
Since `modelJson` was not set for those interfaces, the generated
javascript is invalid.
This commit quickly fixes the problem by setting `modelJson` to the json
`null` value. This is sufficient to make the html2 template happy and
it's the only use for the `modelJson` field.